Afernandez/movvimientos1

+++ src/main/java/com/uisrael/drinkhouse/infraestructura/repositorio/ILoteProductoJpaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.uisrael.drinkhouse.infraestructura.persistencia.jpa.LoteProductoEntity;
 
-public interface ILoteProductoJpaRepositorio extends JpaRepository<LoteProductoEntity, Integer> {
+public interface ILoteProductoJpaRepositorio extends JpaRepository<LoteProductoEntity, Long> {
 
 }
